requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pillow==11.1.0
torch
torchvision
